+++ base.py
@@ -1,0 +1,6 @@
+from flask import Flask
+from flask_sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///project.db'
+db = SQLAlchemy(app)
