+++ models.py
@@ -1,0 +1,21 @@
+from base import db
+from datetime import datetime
+
+
+class User(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(80), nullable=False)
+
+    def __repr__(self):
+        return '<User %r>' % self.name
+
+
+class FridgeItem(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(80), nullable=False)
+    expiry_date = db.Column(db.Date, nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False) # Foreign key to User table
+    input_date = db.Column(db.Date, default=datetime.utcnow)
+
+    def __repr__(self):
+        return '<FridgeItem %r>' % self.name
